requirements.txt
GitPython==2.1.11
PyYAML==3.13